Create Plotting.js file, and move plotting functions from OverviewTab.js. Show all x-label and add slider in the bottom.
