+++ client/src/statistics/OverviewTab.js
@@ -1,120 +1,9 @@
 import React, { useState, useCallback, useEffect } from "react";
 import { Card, Col, Row } from "antd";
-// eslint-disable-next-line import/no-extraneous-dependencies
-import USAMap from "react-usa-map";
-// eslint-disable-next-line import/no-extraneous-dependencies
-import { Column, Pie } from "@ant-design/plots";
 
 import SelectComponent from "./SelectComponent";
 import serverConfig from "../config.json";
-
-function afsByTypePie() {
-  const [data, setData] = useState([]);
-
-  const asyncFetch = () => {
-    fetch(
-      `http://${serverConfig.server_host}:${serverConfig.server_port}/stats/overview/afsByType`
-    )
-      .then((response) => response.json())
-      .then((json) => setData(json))
-      .catch((error) => {
-        // eslint-disable-next-line no-console
-        console.log("fetch data failed", error);
-      });
-  };
-
-  useEffect(() => {
-    asyncFetch();
-  }, []);
-
-  const config = {
-    appendPadding: 10,
-    data,
-    angleField: "numStations",
-    colorField: "fuelType",
-    radius: 0.9,
-    label: {
-      type: "inner",
-      offset: "-30%",
-      content: ({ percent }) => `${(percent * 100).toFixed(0)}%`,
-      style: {
-        fontSize: 14,
-        textAlign: "center",
-      },
-    },
-    interactions: [
-      {
-        type: "element-active",
-      },
-    ],
-  };
-  // eslint-disable-next-line react/jsx-props-no-spreading
-  return <Pie {...config} />;
-}
-
-function afsByStateMap() {
-  const [data, setData] = useState([]);
-
-  const asyncFetch = () => {
-    fetch(
-      `http://${serverConfig.server_host}:${serverConfig.server_port}/stats/overview/afsByState`
-    )
-      .then((response) => response.json())
-      .then((json) => setData(json))
-      .catch((error) => {
-        // eslint-disable-next-line no-console
-        console.log("fetch data failed", error);
-      });
-  };
-
-  useEffect(() => {
-    asyncFetch();
-  }, []);
-
-  const dataDict = {};
-
-  function toDict(item) {
-    dataDict[item.state] = { numStations: item.numStations };
-  }
-  data.forEach(toDict);
-
-  // eslint-disable-next-line react/jsx-props-no-spreading
-  return <USAMap customize={dataDict} />;
-}
-
-function afsByTypeStateBar(data) {
-  const config = {
-    data,
-    xField: "state",
-    yField: "numStations",
-    isGroup: true,
-    isStack: false,
-    seriesField: "stype",
-    columnStyle: {
-      radius: [20, 20, 0, 0],
-    },
-  };
-
-  // eslint-disable-next-line react/jsx-props-no-spreading
-  return <Column {...config} />;
-}
-
-function vehicleByTypeBar(data) {
-  const config = {
-    data,
-    xField: "state",
-    yField: "numVehicle",
-    isGroup: true,
-    isStack: false,
-    seriesField: "vtype",
-    columnStyle: {
-      radius: [20, 20, 0, 0],
-    },
-  };
-
-  // eslint-disable-next-line react/jsx-props-no-spreading
-  return <Column {...config} />;
-}
+import * as plt from "./Plotting";
 
 function OverviewTab() {
   // state variable for station type select menu
@@ -172,11 +61,14 @@ function OverviewTab() {
         <Col span={6}>
           <Card title="🏅 US Alternating fueling resources">
             {" "}
-            {afsByTypePie()}
+            {plt.afsByTypePie()}
           </Card>
         </Col>
         <Col span={18}>
-          <Card title="🏅 Energy Resources Heat map"> {afsByStateMap()}</Card>
+          <Card title="🏅 Energy Resources Heat map">
+            {" "}
+            {plt.afsByStateMap()}
+          </Card>
         </Col>
       </Row>
       <Row gutter={16} type="flex">
@@ -194,7 +86,7 @@ function OverviewTab() {
           </Card>
         </Col>
         <Col span={18}>
-          <Card>Content 2{afsByTypeStateBar(afsByTypeStateData)}</Card>
+          <Card>Content 2{plt.afsByTypeStateBar(afsByTypeStateData)}</Card>
         </Col>
       </Row>
       <Row gutter={16} type="flex">
@@ -212,7 +104,7 @@ function OverviewTab() {
           </Card>
         </Col>
         <Col span={18}>
-          <Card>Content 3{vehicleByTypeBar(vehicleByTypeData)}</Card>
+          <Card>Content 3{plt.vehicleByTypeBar(vehicleByTypeData)}</Card>
         </Col>
       </Row>
     </>
